[master]: Random updates.
Add highlight alias, as it's a pain to run manually.

Add AWS CLI completion, as it doesn't follow the pattern of the rest of
the autocomplete scripts I've got.

Don't try and stop containers on exit when there's nothing to stop, to
suppress an error message.

Reset boot2docker configuration every time, so I don't have to look it
up again.

Broader stash hijacking for SSH, to catch all projects.